+++ native/c/zjb.hpp
@@ -26,6 +26,10 @@ struct ZJob
   std::string full_status;
   std::string retcode;
   std::string correlator;
+  std::string subsystem;
+  std::string type;
+  std::string jobclass;
+  int phase;
 };
 
 struct ZJobDD
@@ -42,7 +46,7 @@ struct ZJobDD
  * @brief Return a list of jobs from an input or default owner
  *
  * @param zjb job returned attributes and error information
- * @param owner_name owner name of the job to query, defaults to currnet user if == "", may use wild cards, i.e.
+ * @param owner_name owner name of the job to query, defaults to current user if == "", may use wild cards, i.e.
  * "IBMUS*"
  * @param jobs populated list returned containing job information array
  * @return int 0 for success; non zero otherwise
@@ -57,13 +61,26 @@ extern "C"
  * @brief Return a list of jobs from an input or default owner
  *
  * @param zjb job returned attributes and error information
- * @param owner_name owner name of the job to query, defaults to currnet user if == "", may use wild cards, i.e.
+ * @param owner_name owner name of the job to query, defaults to current user if == "", may use wild cards, i.e.
  * "IBMUS*"
  * @param prefix job prefix, defaults to "*" if == "", may use wild cards, i.e. "IBMUS*"
  * @param jobs populated list returned containing job information array
  * @return int 0 for success; non zero otherwise
  */
 int zjb_list_by_owner(ZJB *zjb, std::string owner_name, std::string prefix_name, std::vector<ZJob> &jobs);
+
+/**
+ * @brief Return a list of jobs from an input or default owner
+ *
+ * @param zjb job returned attributes and error information
+ * @param owner_name owner name of the job to query, defaults to current user if == "", may use wild cards, i.e.
+ * "IBMUS*"
+ * @param prefix job prefix, defaults to "*" if == "", may use wild cards, i.e. "IBMUS*"
+ * @param status job status, defaults to "*" if == "", supports "ACTIVE" only
+ * @param jobs populated list returned containing job information array
+ * @return int 0 for success; non zero otherwise
+ */
+int zjb_list_by_owner(ZJB *zjb, std::string owner_name, std::string prefix_name, std::string status_name, std::vector<ZJob> &jobs);
 
 /**
  * @brief Return a list of proclib for a job