+++ native/c/zds.hpp
@@ -205,6 +205,16 @@ int zds_create_dsn(ZDS *zds, std::string dsn, DS_ATTRIBUTES attributes, std::str
 int zds_delete_dsn(ZDS *zds, std::string dsn);
 
 /**
+ * @brief Rename a data set
+ *
+ * @param zds data set returned attributes and error information
+ * @param dsn_Before data set name to rename
+ * @param dsn_After new data set name
+ * @return int 0 for success; non zero otherwise
+ */
+int zds_rename_dsn(ZDS *zds, std::string dsn_before, std::string dsn_after);
+
+/**
  * @brief Obtain list of members in a z/OS data set
  *
  * @param zds data set returned attributes and error information